fix: remove controller and move logic directly to Application.kt for pre-course guideline
